refactor : racegame -> racinggame 패키지 이름 변경
